requirements.txt
cupy-cuda12x
pygame